feat(img): add cookies image for love section
This commit introduces new cookies image to the 'trendycookies'
repository. These image are intended to be used in the 'love' 
section of the website.

Summary of Changes:
- Images: Added 'love.jpg' and 'love_2x.jpg' in the 'img' directory.
These images are for standard displays and high-density pixel 
displays respectively.

These new image will enhance the visual appeal of the 'love' 
section, providing users with a better understanding of the products 
offered.