feat(img): add product images for products section
This commit introduces new product images to the 'trendycookies' 
repository. These images are intended to be used in the 'products' 
section of the website.

Summary of Changes:
- Images: Added 'product-7.png' to 'product-12.png' and 
'product-7_2x.png' to 'product-12_2x.png' in the 'img' directory. 
These images are for standard displays and high-density pixel 
displays respectively.

These new images will enhance the visual appeal of the 'products' 
section, providing users with a better understanding of the products 
offered.