feat(img): update website_screenshot.png with the latest website view
This commit updates the 'website_screenshot.png' image in the 'img'
directory. The image is used in the README.md file to provide a visual
representation of the current state of the website.

Summary of Changes:
- Images: Updated 'website_screenshot.png' with the latest screenshot
of the website.

The updated 'website_screenshot.png' now accurately reflects the
current look and feel of the website, providing an up-to-date visual
reference in the README.md file.